requirements_whisperx.txt
whisperx
ctranslate2>=4.5.0
faster-whisper>=1.1.1
nltk>=3.9.1
numpy>=2.1.0
omegaconf>=2.3.0
pandas>=2.2.3
pyannote-audio>=4.0.0
huggingface-hub<1.0.0
torch~=2.8.0
torchaudio~=2.8.0
transformers>=4.48.0
ffmpeg-python
